+++ lib/src/core/common/navigation_pans/restoring_node_pane.dart
@@ -1,7 +1,6 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gui/src/core/extensions/context_extensions.dart';
-import 'package:gui/src/core/utils/daemon_manager/bloc/daemon_cubit.dart';
 import 'package:gui/src/core/utils/gen/localization/locale_keys.dart';
 import 'package:gui/src/features/finish/presentation/screen/finish_page.dart';
 import 'package:gui/src/features/initializing/presentation/screen/initializing_page.dart';
@@ -9,7 +8,7 @@ import 'package:gui/src/features/main/language/core/localization_extension.dart'
 import 'package:gui/src/features/main/navigation_pan_cubit/presentation/cubits/navigation_pan_cubit.dart';
 import 'package:gui/src/features/master_password/presentation/screen/master_password_page.dart';
 import 'package:gui/src/features/restoration_seed/presentation/screen/restoration_seed_page.dart';
-import 'package:gui/src/features/validator_config/presentation/screen/validator_config_page.dart';
+import 'package:gui/src/features/validator_config/presentation/screen/validator_config_screen.dart';
 
 /// ## [RestoringNodePane] Class Documentation
 ///
@@ -20,7 +19,7 @@ import 'package:gui/src/features/validator_config/presentation/screen/validator_
 ///
 /// - **[RestorationSeedPage]**: Restores the wallet using a seed phrase.
 /// - **[MasterPasswordPage]**: Sets or verifies the master password.
-/// - **[ValidatorConfigPage]**: Configures validator settings.
+/// - **[ValidatorConfigScreen]**: Configures validator settings.
 /// - **[InitializingPage]**: Initializes the restored node.
 /// - **[FinishPage]**: Completes the restoration process.
 ///
@@ -97,7 +96,7 @@ class RestoringNodePane extends StatelessWidget {
                     ),
                   ),
                 ),
-                body: ValidatorConfigPage(),
+                body: ValidatorConfigScreen(),
               ),
               PaneItem(
                 icon: const SizedBox(),
@@ -121,14 +120,7 @@ class RestoringNodePane extends StatelessWidget {
                     ),
                   ),
                 ),
-                body: MultiBlocProvider(
-                  providers: [
-                    BlocProvider<DaemonCubit>(
-                      create: (_) => DaemonCubit(),
-                    ),
-                  ],
-                  child: FinishPage(),
-                ),
+                body: FinishPage(),
               ),
             ],
           ),