+++ lib/src/core/common/widgets/custom_password_widget.dart
@@ -65,6 +65,7 @@ class CustomPasswordWidget extends StatelessWidget {
 
   final String placeholder;
   final ValueChanged<String>? onChanged;
+
   final bool readOnly;
   final bool autofocus;
   final TextStyle? textStyle;
@@ -81,12 +82,14 @@ class CustomPasswordWidget extends StatelessWidget {
         child: PasswordBox(
           placeholder: placeholder,
           placeholderStyle: TextStyle(
-            color: AppTheme.of(context).extension<DarkPallet>()!.dark600,
+            color: AppTheme.of(context).extension<DarkPallet>()!.dark900,
           ),
           onChanged: onChanged,
           readOnly: readOnly,
           autofocus: autofocus,
           textAlignVertical: TextAlignVertical.center,
+
+          // TODO(pouria): #71 there is no text style for placeholder .
           style: textStyle ??
               TextStyle(
                 fontSize: 14,