+++ lib/src/core/utils/gen/localization/locale_keys.dart
@@ -4,10 +4,10 @@ abstract class LocaleKeys {
   static const String subtitle = 'subtitle';
   static const String description = 'description';
   static const String switch_language = 'switch_language';
+  static const String applications = 'applications';
   static const String welcome_title = 'welcome_title';
   static const String welcome_description = 'welcome_description';
   static const String start_button_text = 'start_button_text';
-  static const String applications = 'applications';
   static const String initialize_mode = 'initialize_mode';
   static const String wallet_seed = 'wallet_seed';
   static const String confirm_seed = 'confirm_seed';
@@ -16,28 +16,37 @@ abstract class LocaleKeys {
   static const String initializing = 'initializing';
   static const String finish = 'finish';
   static const String initiate_your_node = 'initiate_your_node';
-  static const String initiate_your_node_for_first_time =
-      'initiate_your_node_for_first_time';
-  static const String create_local_node_from_scrache =
-      'create_local_node_from_scrache';
-  static const String restore_local_node_from_seed_pharse =
-      'restore_local_node_from_seed_pharse';
+  static const String initiate_your_node_for_first_time = 'initiate_your_node_for_first_time';
+  static const String create_local_node_from_scrache = 'create_local_node_from_scrache';
+  static const String restore_local_node_from_seed_pharse = 'restore_local_node_from_seed_pharse';
   static const String connect_to_remote_node = 'connect_to_remote_node';
   static const String remote_address = 'remote_address';
   static const String auth_method = 'auth_method';
   static const String username = 'username';
   static const String password = 'password';
   static const String next = 'next';
-  static const String basic_auth = 'basic_auth';
   static const String unlock_wallet = 'unlock_wallet';
   static const String unlock_wallet_description = 'unlock_wallet_description';
   static const String enter_your_password = 'enter_your_password';
   static const String password_cannot_be_empty = 'password_cannot_be_empty';
   static const String incorrect_password = 'incorrect_password';
   static const String unlock_failed = 'unlock_failed';
+  static const String basic_auth = 'basic_auth';
+  static const String error = 'error';
+  static const String select = 'select';
+  static const String working_directory = 'working_directory';
+  static const String choose_your_directory = 'choose_your_directory';
+  static const String select_folder = 'select_folder';
+  static const String validator_config_description = 'validator_config_description';
+  static const String number_of_validators = 'number_of_validators';
+  static const String directory_not_empty = 'directory_not_empty';
+  static const String your_journey_finalized = 'your_journey_finalized';
+  static const String go_to_dashboard = 'go_to_dashboard';
+  static const String go_to_dashboard_description = 'go_to_dashboard_description';
+  static const String your_journey_finalized_description = 'your_journey_finalized_description';
+  static const String your_journey_finalized_description_last = 'your_journey_finalized_description_last';
   static const String generation_seed_title = 'generation_seed_title';
-  static const String generation_seed_description =
-      'generation_seed_description';
+  static const String generation_seed_description = 'generation_seed_description';
   static const String generation_seed_note1 = 'generation_seed_note1';
   static const String generation_seed_note2 = 'generation_seed_note2';
   static const String warning = 'warning';