+++ test/src/core/common/widgets/theme_switcher_test.dart
@@ -54,9 +54,6 @@ void main() {
             );
           },
         ),
-        // child: MaterialApp(
-        //   home: Scaffold(body: child),
-        // ),
       );
     }